+++ shell/Sources/JugnuCore/Permissions/TCCGateError.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public enum TCCGateError: Error, Equatable {
+    case declined(AddonPermission)
+    case openedSettings(AddonPermission)
+}